+++ src/idunno.Bluesky/Properties/Resources.Designer.cs
@@ -88,6 +88,33 @@ namespace idunno.Bluesky.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Post cannot contain both images and gallery images..
+        /// </summary>
+        internal static string PostBuilderCannotHaveImagesAndGalleryImages {
+            get {
+                return ResourceManager.GetString("PostBuilderCannotHaveImagesAndGalleryImages", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Cannot have more than {0} gallery images..
+        /// </summary>
+        internal static string PostBuilderHasTooManyGalleryImages {
+            get {
+                return ResourceManager.GetString("PostBuilderHasTooManyGalleryImages", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Cannot have more than {0} images..
+        /// </summary>
+        internal static string PostBuilderHasTooManyImages {
+            get {
+                return ResourceManager.GetString("PostBuilderHasTooManyImages", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Post cannot have both images and video..
         /// </summary>
         internal static string PostCannotHaveImagesAndVideoValidationError {